refactor(api): move certification center for admin serializer
