+++ api/src/organizational-entities/infrastructure/serializers/jsonapi/certification-center/certification-center-for-admin.serializer.js
@@ -2,7 +2,7 @@ import jsonapiSerializer from 'jsonapi-serializer';
 
 const { Serializer } = jsonapiSerializer;
 
-import { CenterForAdmin } from '../../../../src/certification/enrolment/domain/models/CenterForAdmin.js';
+import { CenterForAdmin } from '../../../../../certification/enrolment/domain/models/CenterForAdmin.js';
 
 const deserialize = function ({ data }) {
   const center = {
